chore(deps): bump deps and address urfave v3 breaking changes

+++ cmd/vela-server/scm.go
@@ -22,7 +22,7 @@ func setupSCM(ctx context.Context, c *cli.Command, tc *tracing.Client) (scm.Serv
 		Address:              c.String("scm.addr"),
 		ClientID:             c.String("scm.client"),
 		ClientSecret:         c.String("scm.secret"),
-		AppID:                c.Int("scm.app.id"),
+		AppID:                c.Int64("scm.app.id"),
 		AppPrivateKey:        c.String("scm.app.private-key"),
 		AppPrivateKeyPath:    c.String("scm.app.private-key.path"),
 		AppPermissions:       c.StringSlice("scm.app.permissions"),
